pyproject.toml
[tool.poetry]
name = "tkursed"
version = "0.0.1"
description = "A completely cursed 2D renderer done entirely with the Tcl/Tk image object."
authors = ["Tony Lechner <tony@tony-lechner.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

